Add files via upload
EYE BLINK DETECTION USING HAAR CASCADE CLASSIFER.
We represent a real time method based on some video and image processing algorithms for eye blink detection. The motivation of this research is the need of disabling who cannot communicate with human. A Haar Cascade Classifier is applied for face and eye detection for getting eye and facial axis information. In addition, the same classifier is used based on Haar- like features to find out the relationship between the eyes and the facial axis for positioning the eyes. An efficient eye tracking method is proposed which uses the position of detected face.
Main objectives:
1]Allow paralysis victims to communicate independently
2]Be accessible to people with financial constraints.

3.3	Hardware Requirements
CPU	: Intel 2.1 GHZ
Memory	: 4GB
Disk	: 100GB
Display	: 15 inch color
Camera Installation: The software needs to detect the blinking of eyes with the assistance of camera. So, the computer should be installed with the camera. Desktop computer doesn’t have camera, so an extra camera needs to be installed and linked to the software. The laptop is equipped with the camera, so it will be linked.

Software (Tools &Technologies) Requirements

Coding	: Python
Platform	: python 3.7 or 3.8 Tool	: Spyder
OS	: Windows 7

Open CV: It is a library of programming functions mainly realtime computer vision. It was originally developed by Intel and it was later supported by Willow Garage and is now maintained by IITSSEZ. OpenCV is written in C++ and is primarily interface is in C++. There are bindings in python, java and MATLAB/Octave.
Python dlib: It is developed by Davis King, the dlib C++ library is a cross-platform package for threading networking, numerical operations, machine learning, computer vision and compression, placing a strong emphasis on extremely high- quality and portable code. From a computer vision perspective, dlib has reveal state-of- the-art implementations, including: Facial landmark detection, Correlation tracking, Deep metric learning.